Moved AlreadyThinking Trigger
Made it so the bot only stops ignoring messages once it has sent its previous message. This was necessary because the new API responds a lot faster than the old one, plus the increased typing time I added, was making it so the bot was responding to a lot of messages that it should've been ignoring.